pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.10"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.3",
    "fastapi[standard]>=0.115.11",
    "greenlet>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
